refactor(api): split routes into dedicated modules based on domain logic
